refactor: remove interceptions API and request logs view
Remove the deprecated /api/v2/aibridge/interceptions endpoint, the
Request Logs frontend page, and all supporting code. The sessions-based
view replaces them. Closes AIGOV-266 and AIGOV-324.

Backend:
- /interceptions HTTP handler and CLI command
- SDK types and client methods for interceptions
- Interceptions list/count SQL queries and search query parser
- dbauthz wrappers and in-memory query implementations

Frontend:
- RequestLogsPage and its row, filter, and view components
- Router route and lazy import for /aibridge/request-logs
- API client method and query function for interceptions
- Mock interception entities

The /models and /clients endpoints stay because the sessions list page
filter menus still need them. The shared filter and icon components
move from RequestLogsPage to AIBridgePage/filters and
AIBridgePage/icons so the sessions pages keep working.
