+++ coderd/database/queries.sql.go
@@ -971,77 +971,6 @@ func (q *sqlQuerier) CalculateAIBridgeInterceptionsTelemetrySummary(ctx context.
 	return i, err
 }
 
-const countAIBridgeInterceptions = `-- name: CountAIBridgeInterceptions :one
-SELECT
-	COUNT(*)
-FROM
-	aibridge_interceptions
-WHERE
-	-- Remove inflight interceptions (ones which lack an ended_at value).
-	aibridge_interceptions.ended_at IS NOT NULL
-	-- Filter by time frame
-	AND CASE
-		WHEN $1::timestamptz != '0001-01-01 00:00:00+00'::timestamptz THEN aibridge_interceptions.started_at >= $1::timestamptz
-		ELSE true
-	END
-	AND CASE
-		WHEN $2::timestamptz != '0001-01-01 00:00:00+00'::timestamptz THEN aibridge_interceptions.started_at <= $2::timestamptz
-		ELSE true
-	END
-	-- Filter initiator_id
-	AND CASE
-		WHEN $3::uuid != '00000000-0000-0000-0000-000000000000'::uuid THEN aibridge_interceptions.initiator_id = $3::uuid
-		ELSE true
-	END
-	-- Filter provider
-	AND CASE
-		WHEN $4::text != '' THEN aibridge_interceptions.provider = $4::text
-		ELSE true
-	END
-	-- Filter provider_name
-	AND CASE
-		WHEN $5::text != '' THEN aibridge_interceptions.provider_name = $5::text
-		ELSE true
-	END
-	-- Filter model
-	AND CASE
-		WHEN $6::text != '' THEN aibridge_interceptions.model = $6::text
-		ELSE true
-	END
-	-- Filter client
-	AND CASE
-		WHEN $7::text != '' THEN COALESCE(aibridge_interceptions.client, 'Unknown') = $7::text
-		ELSE true
-	END
-	-- Authorize Filter clause will be injected below in ListAuthorizedAIBridgeInterceptions
-	-- @authorize_filter
-`
-
-type CountAIBridgeInterceptionsParams struct {
-	StartedAfter  time.Time `db:"started_after" json:"started_after"`
-	StartedBefore time.Time `db:"started_before" json:"started_before"`
-	InitiatorID   uuid.UUID `db:"initiator_id" json:"initiator_id"`
-	Provider      string    `db:"provider" json:"provider"`
-	ProviderName  string    `db:"provider_name" json:"provider_name"`
-	Model         string    `db:"model" json:"model"`
-	Client        string    `db:"client" json:"client"`
-}
-
-func (q *sqlQuerier) CountAIBridgeInterceptions(ctx context.Context, arg CountAIBridgeInterceptionsParams) (int64, error) {
-	row := q.db.QueryRowContext(ctx, countAIBridgeInterceptions,
-		arg.StartedAfter,
-		arg.StartedBefore,
-		arg.InitiatorID,
-		arg.Provider,
-		arg.ProviderName,
-		arg.Model,
-		arg.Client,
-	)
-	var count int64
-	err := row.Scan(&count)
-	return count, err
-}
-
 const countAIBridgeSessions = `-- name: CountAIBridgeSessions :one
 SELECT
 	COUNT(DISTINCT (aibridge_interceptions.session_id, aibridge_interceptions.initiator_id))
@@ -1691,149 +1620,6 @@ func (q *sqlQuerier) ListAIBridgeClients(ctx context.Context, arg ListAIBridgeCl
 			return nil, err
 		}
 		items = append(items, client)
-	}
-	if err := rows.Close(); err != nil {
-		return nil, err
-	}
-	if err := rows.Err(); err != nil {
-		return nil, err
-	}
-	return items, nil
-}
-
-const listAIBridgeInterceptions = `-- name: ListAIBridgeInterceptions :many
-SELECT
-	aibridge_interceptions.id, aibridge_interceptions.initiator_id, aibridge_interceptions.provider, aibridge_interceptions.model, aibridge_interceptions.started_at, aibridge_interceptions.metadata, aibridge_interceptions.ended_at, aibridge_interceptions.api_key_id, aibridge_interceptions.client, aibridge_interceptions.thread_parent_id, aibridge_interceptions.thread_root_id, aibridge_interceptions.client_session_id, aibridge_interceptions.session_id, aibridge_interceptions.provider_name, aibridge_interceptions.credential_kind, aibridge_interceptions.credential_hint,
-	visible_users.id, visible_users.username, visible_users.name, visible_users.avatar_url
-FROM
-	aibridge_interceptions
-JOIN
-	visible_users ON visible_users.id = aibridge_interceptions.initiator_id
-WHERE
-	-- Remove inflight interceptions (ones which lack an ended_at value).
-	aibridge_interceptions.ended_at IS NOT NULL
-	-- Filter by time frame
-	AND CASE
-		WHEN $1::timestamptz != '0001-01-01 00:00:00+00'::timestamptz THEN aibridge_interceptions.started_at >= $1::timestamptz
-		ELSE true
-	END
-	AND CASE
-		WHEN $2::timestamptz != '0001-01-01 00:00:00+00'::timestamptz THEN aibridge_interceptions.started_at <= $2::timestamptz
-		ELSE true
-	END
-	-- Filter initiator_id
-	AND CASE
-		WHEN $3::uuid != '00000000-0000-0000-0000-000000000000'::uuid THEN aibridge_interceptions.initiator_id = $3::uuid
-		ELSE true
-	END
-	-- Filter provider
-	AND CASE
-		WHEN $4::text != '' THEN aibridge_interceptions.provider = $4::text
-		ELSE true
-	END
-	-- Filter provider_name
-	AND CASE
-		WHEN $5::text != '' THEN aibridge_interceptions.provider_name = $5::text
-		ELSE true
-	END
-	-- Filter model
-	AND CASE
-		WHEN $6::text != '' THEN aibridge_interceptions.model = $6::text
-		ELSE true
-	END
-	-- Filter client
-	AND CASE
-		WHEN $7::text != '' THEN COALESCE(aibridge_interceptions.client, 'Unknown') = $7::text
-		ELSE true
-	END
-	-- Cursor pagination
-	AND CASE
-		WHEN $8::uuid != '00000000-0000-0000-0000-000000000000'::uuid THEN (
-			-- The pagination cursor is the last ID of the previous page.
-			-- The query is ordered by the started_at field, so select all
-			-- rows before the cursor and before the after_id UUID.
-			-- This uses a less than operator because we're sorting DESC. The
-			-- "after_id" terminology comes from our pagination parser in
-			-- coderd.
-			(aibridge_interceptions.started_at, aibridge_interceptions.id) < (
-				(SELECT started_at FROM aibridge_interceptions WHERE id = $8),
-				$8::uuid
-			)
-		)
-		ELSE true
-	END
-	-- Authorize Filter clause will be injected below in ListAuthorizedAIBridgeInterceptions
-	-- @authorize_filter
-ORDER BY
-	aibridge_interceptions.started_at DESC,
-	aibridge_interceptions.id DESC
-LIMIT COALESCE(NULLIF($10::integer, 0), 100)
-OFFSET $9
-`
-
-type ListAIBridgeInterceptionsParams struct {
-	StartedAfter  time.Time `db:"started_after" json:"started_after"`
-	StartedBefore time.Time `db:"started_before" json:"started_before"`
-	InitiatorID   uuid.UUID `db:"initiator_id" json:"initiator_id"`
-	Provider      string    `db:"provider" json:"provider"`
-	ProviderName  string    `db:"provider_name" json:"provider_name"`
-	Model         string    `db:"model" json:"model"`
-	Client        string    `db:"client" json:"client"`
-	AfterID       uuid.UUID `db:"after_id" json:"after_id"`
-	Offset        int32     `db:"offset_" json:"offset_"`
-	Limit         int32     `db:"limit_" json:"limit_"`
-}
-
-type ListAIBridgeInterceptionsRow struct {
-	AIBridgeInterception AIBridgeInterception `db:"aibridge_interception" json:"aibridge_interception"`
-	VisibleUser          VisibleUser          `db:"visible_user" json:"visible_user"`
-}
-
-func (q *sqlQuerier) ListAIBridgeInterceptions(ctx context.Context, arg ListAIBridgeInterceptionsParams) ([]ListAIBridgeInterceptionsRow, error) {
-	rows, err := q.db.QueryContext(ctx, listAIBridgeInterceptions,
-		arg.StartedAfter,
-		arg.StartedBefore,
-		arg.InitiatorID,
-		arg.Provider,
-		arg.ProviderName,
-		arg.Model,
-		arg.Client,
-		arg.AfterID,
-		arg.Offset,
-		arg.Limit,
-	)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	var items []ListAIBridgeInterceptionsRow
-	for rows.Next() {
-		var i ListAIBridgeInterceptionsRow
-		if err := rows.Scan(
-			&i.AIBridgeInterception.ID,
-			&i.AIBridgeInterception.InitiatorID,
-			&i.AIBridgeInterception.Provider,
-			&i.AIBridgeInterception.Model,
-			&i.AIBridgeInterception.StartedAt,
-			&i.AIBridgeInterception.Metadata,
-			&i.AIBridgeInterception.EndedAt,
-			&i.AIBridgeInterception.APIKeyID,
-			&i.AIBridgeInterception.Client,
-			&i.AIBridgeInterception.ThreadParentID,
-			&i.AIBridgeInterception.ThreadRootID,
-			&i.AIBridgeInterception.ClientSessionID,
-			&i.AIBridgeInterception.SessionID,
-			&i.AIBridgeInterception.ProviderName,
-			&i.AIBridgeInterception.CredentialKind,
-			&i.AIBridgeInterception.CredentialHint,
-			&i.VisibleUser.ID,
-			&i.VisibleUser.Username,
-			&i.VisibleUser.Name,
-			&i.VisibleUser.AvatarURL,
-		); err != nil {
-			return nil, err
-		}
-		items = append(items, i)
 	}
 	if err := rows.Close(); err != nil {
 		return nil, err
