+++ coderd/database/queries/aibridge.sql
@@ -113,122 +113,6 @@ ORDER BY
 	created_at ASC,
 	id ASC;
 
--- name: CountAIBridgeInterceptions :one
-SELECT
-	COUNT(*)
-FROM
-	aibridge_interceptions
-WHERE
-	-- Remove inflight interceptions (ones which lack an ended_at value).
-	aibridge_interceptions.ended_at IS NOT NULL
-	-- Filter by time frame
-	AND CASE
-		WHEN @started_after::timestamptz != '0001-01-01 00:00:00+00'::timestamptz THEN aibridge_interceptions.started_at >= @started_after::timestamptz
-		ELSE true
-	END
-	AND CASE
-		WHEN @started_before::timestamptz != '0001-01-01 00:00:00+00'::timestamptz THEN aibridge_interceptions.started_at <= @started_before::timestamptz
-		ELSE true
-	END
-	-- Filter initiator_id
-	AND CASE
-		WHEN @initiator_id::uuid != '00000000-0000-0000-0000-000000000000'::uuid THEN aibridge_interceptions.initiator_id = @initiator_id::uuid
-		ELSE true
-	END
-	-- Filter provider
-	AND CASE
-		WHEN @provider::text != '' THEN aibridge_interceptions.provider = @provider::text
-		ELSE true
-	END
-	-- Filter provider_name
-	AND CASE
-		WHEN @provider_name::text != '' THEN aibridge_interceptions.provider_name = @provider_name::text
-		ELSE true
-	END
-	-- Filter model
-	AND CASE
-		WHEN @model::text != '' THEN aibridge_interceptions.model = @model::text
-		ELSE true
-	END
-	-- Filter client
-	AND CASE
-		WHEN @client::text != '' THEN COALESCE(aibridge_interceptions.client, 'Unknown') = @client::text
-		ELSE true
-	END
-	-- Authorize Filter clause will be injected below in ListAuthorizedAIBridgeInterceptions
-	-- @authorize_filter
-;
-
--- name: ListAIBridgeInterceptions :many
-SELECT
-	sqlc.embed(aibridge_interceptions),
-	sqlc.embed(visible_users)
-FROM
-	aibridge_interceptions
-JOIN
-	visible_users ON visible_users.id = aibridge_interceptions.initiator_id
-WHERE
-	-- Remove inflight interceptions (ones which lack an ended_at value).
-	aibridge_interceptions.ended_at IS NOT NULL
-	-- Filter by time frame
-	AND CASE
-		WHEN @started_after::timestamptz != '0001-01-01 00:00:00+00'::timestamptz THEN aibridge_interceptions.started_at >= @started_after::timestamptz
-		ELSE true
-	END
-	AND CASE
-		WHEN @started_before::timestamptz != '0001-01-01 00:00:00+00'::timestamptz THEN aibridge_interceptions.started_at <= @started_before::timestamptz
-		ELSE true
-	END
-	-- Filter initiator_id
-	AND CASE
-		WHEN @initiator_id::uuid != '00000000-0000-0000-0000-000000000000'::uuid THEN aibridge_interceptions.initiator_id = @initiator_id::uuid
-		ELSE true
-	END
-	-- Filter provider
-	AND CASE
-		WHEN @provider::text != '' THEN aibridge_interceptions.provider = @provider::text
-		ELSE true
-	END
-	-- Filter provider_name
-	AND CASE
-		WHEN @provider_name::text != '' THEN aibridge_interceptions.provider_name = @provider_name::text
-		ELSE true
-	END
-	-- Filter model
-	AND CASE
-		WHEN @model::text != '' THEN aibridge_interceptions.model = @model::text
-		ELSE true
-	END
-	-- Filter client
-	AND CASE
-		WHEN @client::text != '' THEN COALESCE(aibridge_interceptions.client, 'Unknown') = @client::text
-		ELSE true
-	END
-	-- Cursor pagination
-	AND CASE
-		WHEN @after_id::uuid != '00000000-0000-0000-0000-000000000000'::uuid THEN (
-			-- The pagination cursor is the last ID of the previous page.
-			-- The query is ordered by the started_at field, so select all
-			-- rows before the cursor and before the after_id UUID.
-			-- This uses a less than operator because we're sorting DESC. The
-			-- "after_id" terminology comes from our pagination parser in
-			-- coderd.
-			(aibridge_interceptions.started_at, aibridge_interceptions.id) < (
-				(SELECT started_at FROM aibridge_interceptions WHERE id = @after_id),
-				@after_id::uuid
-			)
-		)
-		ELSE true
-	END
-	-- Authorize Filter clause will be injected below in ListAuthorizedAIBridgeInterceptions
-	-- @authorize_filter
-ORDER BY
-	aibridge_interceptions.started_at DESC,
-	aibridge_interceptions.id DESC
-LIMIT COALESCE(NULLIF(@limit_::integer, 0), 100)
-OFFSET @offset_
-;
-
 -- name: ListAIBridgeTokenUsagesByInterceptionIDs :many
 SELECT
 	*
