--- docs/reference/cli/aibridge.md
@@ -1,16 +1,0 @@
-<!-- DO NOT EDIT | GENERATED CONTENT -->
-# aibridge
-
-Manage AI Bridge.
-
-## Usage
-
-```console
-coder aibridge
-```
-
-## Subcommands
-
-| Name                                                      | Purpose                         |
-|-----------------------------------------------------------|---------------------------------|
-| [<code>interceptions</code>](./aibridge_interceptions.md) | Manage AI Bridge interceptions. |
