--- docs/reference/cli/aibridge_interceptions.md
@@ -1,16 +1,0 @@
-<!-- DO NOT EDIT | GENERATED CONTENT -->
-# aibridge interceptions
-
-Manage AI Bridge interceptions.
-
-## Usage
-
-```console
-coder aibridge interceptions
-```
-
-## Subcommands
-
-| Name                                                  | Purpose                               |
-|-------------------------------------------------------|---------------------------------------|
-| [<code>list</code>](./aibridge_interceptions_list.md) | List AI Bridge interceptions as JSON. |
