+++ enterprise/cli/testdata/coder_--help.golden
@@ -16,7 +16,6 @@ USAGE:
 SUBCOMMANDS:
     agent-firewall         Network isolation tool for monitoring and restricting
                            HTTP/HTTPS requests
-    aibridge               Manage AI Bridge.
     external-workspaces    Create or manage external workspaces
     features               List Enterprise features
     groups                 Manage groups
