--- enterprise/cli/testdata/coder_aibridge_--help.golden
@@ -1,12 +1,0 @@
-coder v0.0.0-devel
-
-USAGE:
-  coder aibridge
-
-  Manage AI Bridge.
-
-SUBCOMMANDS:
-    interceptions    Manage AI Bridge interceptions.
-
-———
-Run `coder --help` for a list of global options.
