+++ site/src/api/queries/aiBridge.ts
@@ -1,7 +1,6 @@
 import type { UseInfiniteQueryOptions } from "react-query";
 import { API } from "#/api/api";
 import type {
-	AIBridgeListInterceptionsResponse,
 	AIBridgeListSessionsResponse,
 	AIBridgeSessionThreadsResponse,
 } from "#/api/typesGenerated";
@@ -9,24 +8,6 @@ import { useFilterParamsKey } from "#/components/Filter/Filter";
 import type { UsePaginatedQueryOptions } from "#/hooks/usePaginatedQuery";
 
 const SESSION_THREADS_INFINITE_PAGE_SIZE = 20;
-
-export const paginatedInterceptions = (
-	searchParams: URLSearchParams,
-): UsePaginatedQueryOptions<AIBridgeListInterceptionsResponse, string> => {
-	return {
-		searchParams,
-		queryPayload: () => searchParams.get(useFilterParamsKey) ?? "",
-		queryKey: ({ limit, offset, payload }) => {
-			return ["aiBridgeInterceptions", limit, offset, payload] as const;
-		},
-		queryFn: ({ limit, offset, payload }) =>
-			API.getAIBridgeInterceptions({
-				offset,
-				limit,
-				q: payload,
-			}),
-	};
-};
 
 export const paginatedSessions = (
 	searchParams: URLSearchParams,
