+++ site/src/pages/AIBridgePage/icons/AIBridgeModelIcon.tsx
