+++ site/src/testHelpers/entities.ts
@@ -5417,81 +5417,6 @@ export const MockDisplayNameTasks = [
 	},
 ] satisfies TypesGen.Task[];
 
-export const MockInterception: TypesGen.AIBridgeInterception = {
-	id: "5c1da48a-9eb0-440e-9c82-5bc5692a603d",
-	initiator: {
-		id: "1ebb7622-e6ea-45b4-b244-dda30afc7238",
-		username: "testuser",
-		avatar_url: "https://example.com/avatar.png",
-	},
-	provider: "openai",
-	provider_name: "openai",
-	model: "gpt-4o",
-	started_at: "2022-05-17T17:39:01.382927298Z",
-	ended_at: "2022-05-17T17:39:01.382927298Z",
-	token_usages: [
-		{
-			id: "32e7fd17-24be-46b9-b867-2f0adfd42aff",
-			interception_id: "5c1da48a-9eb0-440e-9c82-5bc5692a603d",
-			provider_response_id: "res_1234567890",
-			input_tokens: 5,
-			output_tokens: 1,
-			cache_read_input_tokens: 3,
-			cache_write_input_tokens: 1,
-			metadata: {},
-			created_at: "2022-05-17T17:39:01.382927298Z",
-		},
-	],
-	metadata: {},
-	user_prompts: [
-		{
-			id: "85154044-818e-4ee4-bac2-87f3ac8f066b",
-			interception_id: "5c1da48a-9eb0-440e-9c82-5bc5692a603d",
-			provider_response_id: "res_1234567890",
-			prompt: "Hello OpenAI",
-			metadata: {},
-			created_at: "2022-05-17T17:39:01.382927298Z",
-		},
-	],
-	tool_usages: [],
-	api_key_id: "5c1da48a-9eb0-440e-9c82-5bc5692a603d",
-	client: "Claude Code",
-};
-
-export const MockInterceptionAnthropic: TypesGen.AIBridgeInterception = {
-	...MockInterception,
-	id: "e5610f5b-2d6c-43db-b1c0-1dfcc6531f04",
-	provider: "anthropic",
-	model: "claude-sonnet-4.5",
-	user_prompts: [
-		{
-			id: "c820f31f-0170-4044-8b7c-b1b18747b4fb",
-			interception_id: "e5610f5b-2d6c-43db-b1c0-1dfcc6531f04",
-			provider_response_id: "res_2345678901",
-			prompt: "Hello Anthropic",
-			metadata: {},
-			created_at: "2022-05-17T17:39:01.382927298Z",
-		},
-	],
-};
-
-export const MockInterceptionCopilot: TypesGen.AIBridgeInterception = {
-	...MockInterception,
-	id: "22c9d31e-1a1f-464a-b397-562958599aa8",
-	provider: "copilot",
-	model: "claude-opus-4-5",
-	user_prompts: [
-		{
-			id: "c6c613d1-177e-416f-95b5-c7f0eeefb922",
-			interception_id: "22c9d31e-1a1f-464a-b397-562958599aa8",
-			provider_response_id: "res_3456789012",
-			prompt: "Hello Copilot",
-			metadata: {},
-			created_at: "2022-05-17T17:39:01.382927298Z",
-		},
-	],
-};
-
 export const MockSession: TypesGen.AIBridgeSession = {
 	id: "c8f2df8c-149c-43e1-9d51-898daaa2c505",
 	initiator: {
